Fix uploads no file problem.
The backend was not checking that a file was actually
present. This caused an exception. The photo construction
logic in the model now checks that the image is actually
present before trying to set the photo properties. If the
photo is no present then it results in an invalid photo
object and the user is redirected back to the new photo
form.